OpenWeather - Weather: WeatherDetail 구조체의 특정 프로퍼티의 타입 수정 - WeatherViewModel: 현재 장소의 날씨정보 추가 표시에 대한 output용 Observable 타입의 프로퍼티 추가 생성 및 inputCurrentWeather bind시 값을 가지도록 설정 - Enum: WeatherDetailType 열거형을 생성하여 현재 장소의 날씨정보 추가 표시할 데이터들을 구분하여 설정 - MainViewController: 현재 장소의 날씨정보를 collectionView로 추가로 확인할 수 있도록 생성 및 오토레이아웃 설정 / CollectionView들의 layout을 UICollectionViewDelegateFlowLayout를 통해 extension하여 설정 & UICollectionViewLayout+Extension 파일 삭제 - WeatherDetailCollectionViewCell: WeatherDetailCollectionView에 들어갈 Cell UI 메서드 생성 및 오토레이아웃 설정 - etc: 특정 뷰 오토레이아웃 수정 / 폴더링 작업 / alert 함수를 extension 파일에 따로 넣어줌

+++ OpenWeather/Model/Weather.swift
@@ -98,8 +98,8 @@ struct WeatherDetail: Decodable {
     let feels_like: Double
     let temp_min: Double
     let temp_max: Double
-    let pressure: Double
-    let humidity: Double
+    let pressure: Int
+    let humidity: Int
     
     var tempStr: String {
         "\(self.temp)" + "º"
